pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiosqlite>=0.22.1",
    "numpy>=2.4.0",
    "pillow>=12.1.0",
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=22.5",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "ultralytics>=8.3.250",
]
